face-recognition-service/requirements-deploy.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.21.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
facenet-pytorch==2.5.2
Pillow==9.5.0
--find-links https://download.pytorch.org/whl/torch_stable.html